refactor(layout): MobileSidebar sous 300 lignes — éclatement complet (H12)
- useMobileSidebar.js(props, emit): détection mobile (ferme si desktop),
  items nav par rôle (secondaire + admin), chemin actif, déconnexion
- MobileSidebarNav: sections Navigation/Administration/Compte présentationnelles
  + CSS nav déplacé verbatim; émet close/logout
- MobileSidebar.vue: orchestrateur 207 lignes (400 → 207), chrome (overlay,
  header, footer, transitions, scrollbar) conservé
- Tests: useMobileSidebar (logique) + montage MobileSidebarNav; MobileSidebar.test conservé
